[feat] add arrow and some refactoring
